requirements.txt
findspark>=2.0.1
pyspark>=3.5.1
paho-mqtt>=1.6.1
pandas>=1.5.3
joblib>=1.3.1
faker>=18.9.0
azure-storage-blob>=12.19.1
scikit-learn==1.6.1